New API

+++ xFunc.Maths/Expressions/Trigonometric/Tan.cs
@@ -65,7 +65,7 @@ namespace xFunc.Maths.Expressions.Trigonometric
         /// <seealso cref="ExpressionParameters" />
         protected override double CalculateDergee(ExpressionParameters parameters)
         {
-            var radian = argument.Calculate(parameters) * Math.PI / 180;
+            var radian = (double)argument.Calculate(parameters) * Math.PI / 180;
 
             return Math.Tan(radian);
         }
@@ -80,7 +80,7 @@ namespace xFunc.Maths.Expressions.Trigonometric
         /// <seealso cref="ExpressionParameters" />
         protected override double CalculateRadian(ExpressionParameters parameters)
         {
-            return Math.Tan(argument.Calculate(parameters));
+            return Math.Tan((double)argument.Calculate(parameters));
         }
 
         /// <summary>
@@ -93,7 +93,7 @@ namespace xFunc.Maths.Expressions.Trigonometric
         /// <seealso cref="ExpressionParameters" />
         protected override double CalculateGradian(ExpressionParameters parameters)
         {
-            var radian = argument.Calculate(parameters) * Math.PI / 200;
+            var radian = (double)argument.Calculate(parameters) * Math.PI / 200;
 
             return Math.Tan(radian);
         }
